RyggRuneDetector/requirements.txt
django>=3.2.14
Pillow 
numpy 
onnxruntime
opencv-python
torch
torchvision
gunicorn
whitenoise